+++ src/panels/ActionsPanel/ActionsPanel.tsx
@@ -1,122 +1,22 @@
-import { Button, Flex, Grid, ScrollArea } from '@mantine/core';
+import {
+  ActionIcon,
+  Button,
+  Container,
+  Flex,
+  Grid,
+  Group,
+  ScrollArea
+} from '@mantine/core';
 
 import { FilterList } from '@/components/FilterList/FilterList';
 import { generateMatcherFunctionByKeys } from '@/components/FilterList/util';
-import { FolderIcon } from '@/icons/icons';
+import { BackArrowIcon, FolderIcon } from '@/icons/icons';
 import { setActionsPath } from '@/redux/actions/actionsSlice';
 import { useAppDispatch, useAppSelector } from '@/redux/hooks';
-import { Action } from '@/types/types';
+import { Action, FolderContent } from '@/types/types';
 
 import { ActionsButton } from './ActionsButton';
-
-interface FolderContent {
-  actions: Action[];
-  folders: Folder;
-}
-
-interface Folder {
-  [key: string]: FolderContent;
-}
-
-function actionsForLevel(
-  actions: Action[],
-  navigationPath: string,
-  isInitialized: boolean
-) {
-  const actionsMapped: Folder = { '/': { actions: [], folders: {} } };
-  if (!isInitialized) {
-    return actionsMapped['/'];
-  }
-  actions.forEach((action) => {
-    // If there is no backslash at beginning of GUI path, add that manually
-    // (there should always be though)
-    let actionGuiPath = action.guiPath;
-    if (action.guiPath.length > 0 && action.guiPath[0] !== '/') {
-      actionGuiPath = `/${action.guiPath}`;
-    }
-
-    let guiFolders = actionGuiPath.split('/');
-    // Remove all empty strings: which is what we get before initial slash and
-    // if the path is just a slash
-    guiFolders = guiFolders.filter((s) => s !== '');
-
-    // Add to top level actions (no gui path)
-    if (guiFolders.length === 0) {
-      actionsMapped['/'].actions.push(action);
-    }
-
-    // Add actions of other levels
-    let parent = actionsMapped['/'];
-    while (guiFolders.length > 0) {
-      const folderName = guiFolders.shift();
-      if (folderName === undefined) {
-        throw 'Foldername was undefined! Typescript conversion need some look over';
-      }
-      if (!(folderName in parent.folders)) {
-        parent.folders[folderName] = { actions: [], folders: {} };
-      }
-      if (guiFolders.length === 0) {
-        parent.folders[folderName].actions.push(action);
-      } else {
-        parent = parent.folders[folderName];
-      }
-    }
-  });
-
-  const navPath = navigationPath;
-  let actionsForPath = actionsMapped['/'];
-  if (navPath.length > 1) {
-    const folders = navPath.split('/');
-    folders.shift();
-    while (folders.length > 0) {
-      const folderName = folders.shift();
-      if (folderName === undefined) {
-        throw 'Foldername was undefined! Typescript conversion need some look over';
-      }
-      actionsForPath = actionsForPath.folders[folderName];
-    }
-  }
-  return actionsForPath;
-}
-
-// Truncate navigation path if too long
-function truncatePath(navigationPath: string) {
-  const NAVPATH_LENGTH_LIMIT = 60;
-  const shouldTruncate = navigationPath.length > NAVPATH_LENGTH_LIMIT;
-
-  let truncatedPath = navigationPath;
-  if (shouldTruncate) {
-    let originalPath = navigationPath;
-    if (originalPath[0] === '/') {
-      originalPath = originalPath.substring(1);
-    }
-    const pieces = originalPath.split('/');
-    if (pieces.length > 1) {
-      // TODO: maybe keep more pieces of the path, if possible?
-      truncatedPath = `/${pieces[0]}/.../${pieces[pieces.length - 1]}`;
-    } else {
-      truncatedPath = navigationPath.substring(0, NAVPATH_LENGTH_LIMIT);
-    }
-  }
-
-  return truncatedPath;
-}
-
-function getDisplayedActions(allActions: Action[], navPath: string) {
-  return allActions.filter((action) => {
-    if (navPath.length === 1) {
-      return true;
-    }
-    const navPathGui = navPath.split('/');
-    const actionPathGui = action.guiPath?.split('/');
-    for (let i = 0; i < navPathGui.length; i++) {
-      if (actionPathGui?.[i] !== navPathGui[i]) {
-        return false;
-      }
-    }
-    return true;
-  });
-}
+import { actionsForLevel, getDisplayedActions, truncatePath } from './util';
 
 export function ActionsPanel() {
   const allActions = useAppSelector((state) => state.actions.actions);
@@ -129,34 +29,29 @@ export function ActionsPanel() {
   const displayedNavigationPath = truncatePath(navigationPath);
   const displayedActions = getDisplayedActions(allActions, navigationPath);
 
-  function actionPath(action: string) {
-    dispatch(setActionsPath(action));
-  }
-
-  function addNavPath(path: string) {
+  function addNavPath(path: string): void {
     let navString = navigationPath;
     if (navigationPath === '/') {
       navString += path;
     } else {
       navString += `/${path}`;
     }
-    actionPath(navString);
+    dispatch(setActionsPath(navString));
   }
 
-  function goBack() {
+  function goBack(): void {
     let navString = navigationPath;
     navString = navString.substring(0, navString.lastIndexOf('/'));
     if (navString.length === 0) {
       navString = '/';
     }
-    actionPath(navString);
+    dispatch(setActionsPath(navString));
   }
 
-  function folderButton(key: string) {
+  function folderButton(key: string): React.JSX.Element {
     return (
-      <Button onClick={() => addNavPath(key)} key={key}>
-        <FolderIcon />
-        <p>{key}</p>
+      <Button leftSection={<FolderIcon />} onClick={() => addNavPath(key)} key={key}>
+        {key}
       </Button>
     );
   }
@@ -177,15 +72,15 @@ export function ActionsPanel() {
     ));
   }
 
-  function getBackButton() {
+  function displayBackButton(): React.JSX.Element {
     if (navigationPath !== '/') {
       return (
-        <Button onClick={goBack} key={'backbtn'}>
-          Back
-        </Button>
+        <ActionIcon onClick={goBack} key={'backbtn'}>
+          <BackArrowIcon />
+        </ActionIcon>
       );
     }
-    return null;
+    return <></>;
   }
 
   if (actionLevel === undefined) {
@@ -196,29 +91,30 @@ export function ActionsPanel() {
   const isEmpty = actionLevel === null;
   const actions = isEmpty ? <div>No Actions</div> : getActionContent(actionLevel);
   const folders = isEmpty ? <div>No Folders</div> : getFoldersContent(actionLevel);
-  const backButton = getBackButton();
   return (
-    <ScrollArea h={'100%'}>
-      <Flex>
-        {backButton}
-        <p>{`${displayedNavigationPath}`}</p>
-      </Flex>
-      <FilterList>
-        <FilterList.InputField placeHolderSearchText={'Search for an action...'} />
-        <FilterList.Favorites>
-          <Grid>
-            {folders}
-            {actions}
-          </Grid>
-        </FilterList.Favorites>
-        <FilterList.Data<Action>
-          data={displayedActions}
-          renderElement={(action: Action) => (
-            <ActionsButton key={`${action.identifier}Filtered`} action={action} />
-          )}
-          matcherFunc={generateMatcherFunctionByKeys(['identifier', 'name', 'guiPath'])}
-        />
-      </FilterList>
-    </ScrollArea>
+    <Container>
+      <ScrollArea h={'100%'}>
+        <Group gap={'xs'}>
+          {displayBackButton()}
+          <p>{`${displayedNavigationPath}`}</p>
+        </Group>
+        <FilterList>
+          <FilterList.InputField placeHolderSearchText={'Search for an action...'} />
+          <FilterList.Favorites>
+            <Grid>
+              {folders}
+              {actions}
+            </Grid>
+          </FilterList.Favorites>
+          <FilterList.Data<Action>
+            data={displayedActions}
+            renderElement={(action: Action) => (
+              <ActionsButton key={`${action.identifier}Filtered`} action={action} />
+            )}
+            matcherFunc={generateMatcherFunctionByKeys(['identifier', 'name', 'guiPath'])}
+          />
+        </FilterList>
+      </ScrollArea>
+    </Container>
   );
 }
